Moved input to Form folder
